# Conflicts:
#	src/KurrentDB.SecondaryIndexing/Indexes/Default/DefaultIndex.cs
#	src/KurrentDB.SecondaryIndexing/Indexes/Default/DefaultIndexProcessor.cs
#	src/KurrentDB.SecondaryIndexing/Indexes/EventType/EventTypeIndexProcessor.cs
#	src/KurrentDB.SecondaryIndexing/Indexes/EventType/EventTypeSql.cs
#	src/KurrentDB.SecondaryIndexing/Indexes/Stream/StreamIndexProcessor.cs
#	src/KurrentDB.SecondaryIndexing/KurrentDB.SecondaryIndexing.csproj
#	src/KurrentDB.SecondaryIndexing/Storage/1_Schema.sql
#	src/KurrentDB.SecondaryIndexing/Storage/DuckDbExtensions.cs

+++ src/KurrentDB.SecondaryIndexing/Storage/1_Schema.sql
@@ -20,7 +20,7 @@ create table if not exists streams (
 
 create table if not exists idx_all (
 	seq bigint not null,
-	event_number int4 not null,
+	event_number bigint not null,
 	log_position bigint not null,
 	created bigint not null,
 	stream bigint not null,
